refactor: data layer 패키지 분리
